requirements.txt
requests
-e .
py-stackexchange >= 2.0